git-worklog: fix bug when parsing multi-word arguments
this patch fixes an issue when parsing arguments that caused arguments
with spaces in them to be incorrectly separated and assigned to
START_DATE and END_DATE; causing the real second argument to become a
third, and throw an error due to reaching the maximum argument limit. by
changing the Input Field Separator, we are able to reliably parse
multi-word arguments (like 'last friday').