remove quotes from var containing command
even though shellcheck warns about this, we need to remove
the quotes surrounding the variable here, because the interpreter will
attempt to execute it as a single command, e.g. all options and
arguments are included in the command search

we are only storing the command in a variable for re-use due to how
debugging/dry-run is implemented. that should probably be changed in a
future commit.